requirements.txt
altgraph==0.10.2
asn1crypto==0.23.0
attrs==17.2.0
Automat==0.6.0
bdist-mpkg==0.5.0
bonjour-py==0.3
cffi==1.11.2
confusable-homoglyphs==2.0.2
constantly==15.1.0
cryptography==2.1.2
cssselect==1.0.1
Django==1.10.4
django-bootstrap-breadcrumbs==0.9.0
django-cors-headers==2.1.0
django-crispy-forms==1.7.0
django-crumbs==0.5.0
django-filter==1.1.0
django-markdown-deux==1.0.5
django-mongodb-engine==0.6.0
django-pagedown==0.1.3
django-registration==2.3
django-vanilla-views==1.0.4
django-versatileimagefield==1.8.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djangotoolbox==1.8.0
enum34==1.1.6
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipaddress==1.0.18
lxml==4.1.0
macholib==1.5.1
Markdown==2.6.10
markdown2==2.3.5
matplotlib==1.3.1
modulegraph==0.10.4
mongoengine==0.14.3
nltk==3.2.5
numpy==1.8.0rc1
olefile==0.44
parsel==1.2.0
Pillow==4.3.0
psycopg2==2.7.3.1
py2app==0.7.3
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
PyJWT==1.5.3
pymongo==3.5.1
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==17.3.0
pyparsing==2.0.1
python-dateutil==1.5
pytz==2013.7
queuelib==1.4.2
scipy==0.13.0b1
Scrapy==1.4.0
scrapy-jsonrpc==0.3.0
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
vboxapi==1.0
virtualenv==15.1.0
w3lib==1.18.0
xattr==0.6.4
zope.interface==4.4.3
